+++ tests/testthat/test-local-error-checking.R
@@ -34,3 +34,21 @@ test_that("parameter",{
   purrr::walk(prm_funs, ~expect_silent(purrr::invoke(.x, name = "test123")))
 
 })
+
+# observation -------------------------------------------------------------
+
+test_that("observation",{
+  obs_funs <- lsf.str("package:assemblerr", pattern = "^obs_")
+
+  # declaration
+  purrr::walk(obs_funs, ~expect_error(purrr::invoke(.x, prediction = "a+a")))
+  purrr::walk(obs_funs, ~expect_error(purrr::invoke(.x, prediction = quote(a+a))))
+  purrr::walk(obs_funs, ~expect_silent(purrr::invoke(.x, prediction = "test123")))
+  purrr::walk(obs_funs, ~expect_silent(purrr::invoke(.x, prediction = ~test)))
+  purrr::walk(obs_funs, ~expect_silent(purrr::invoke(.x, prediction = ~C["central"])))
+
+  # name
+  purrr::walk(obs_funs, ~expect_error(purrr::invoke(.x, prediction = ~test, name = "_test")))
+  purrr::walk(obs_funs, ~expect_silent(purrr::invoke(.x, prediction = ~test, name = "conc")))
+
+})